Group function handlers like in Vim help
See `:help function-list`
