+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/vimscript/model/functions/handlers/floatFunctions/AbsFunctionHandler.kt
@@ -1,27 +1,28 @@
 /*
- * Copyright 2003-2023 The IdeaVim authors
+ * Copyright 2003-2025 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
  * https://opensource.org/licenses/MIT.
  */
 
-package com.maddyhome.idea.vim.vimscript.model.functions.handlers
+package com.maddyhome.idea.vim.vimscript.model.functions.handlers.floatFunctions
 
 import com.intellij.vim.annotations.VimscriptFunction
 import com.maddyhome.idea.vim.api.ExecutionContext
 import com.maddyhome.idea.vim.api.VimEditor
 import com.maddyhome.idea.vim.vimscript.model.VimLContext
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimDataType
-import com.maddyhome.idea.vim.vimscript.model.datatypes.VimString
+import com.maddyhome.idea.vim.vimscript.model.datatypes.VimFloat
+import com.maddyhome.idea.vim.vimscript.model.datatypes.VimInt
 import com.maddyhome.idea.vim.vimscript.model.expressions.Expression
 import com.maddyhome.idea.vim.vimscript.model.functions.FunctionHandler
-import java.util.*
+import kotlin.math.abs
 
-@VimscriptFunction(name = "toupper")
-internal class ToupperFunctionHandler : FunctionHandler() {
-  override val minimumNumberOfArguments: Int = 1
-  override val maximumNumberOfArguments: Int = 1
+@VimscriptFunction(name = "abs")
+internal class AbsFunctionHandler : FunctionHandler() {
+  override val minimumNumberOfArguments = 1
+  override val maximumNumberOfArguments = 1
 
   override fun doFunction(
     argumentValues: List<Expression>,
@@ -29,7 +30,11 @@ internal class ToupperFunctionHandler : FunctionHandler() {
     context: ExecutionContext,
     vimContext: VimLContext,
   ): VimDataType {
-    val argumentString = argumentValues[0].evaluate(editor, context, vimContext).asString()
-    return VimString(argumentString.uppercase(Locale.getDefault()))
+    val argument = argumentValues[0].evaluate(editor, context, vimContext)
+    return if (argument is VimFloat) {
+      VimFloat(abs(argument.value))
+    } else {
+      VimInt(abs(argument.asDouble().toInt()))
+    }
   }
 }
